+++ hypervisor/lib.c
@@ -11,9 +11,8 @@
  */
 
 #include <jailhouse/string.h>
-#include <jailhouse/types.h>
 
-void *memset(void *s, int c, unsigned long n)
+void *memset(void *s, int c, size_t n)
 {
 	u8 *p = s;
 
@@ -33,7 +32,7 @@ int strcmp(const char *s1, const char *s2)
 	return *(unsigned char *)s1 - *(unsigned char *)s2;
 }
 
-void *memcpy(void *dest, const void *src, unsigned long n)
+void *memcpy(void *dest, const void *src, size_t n)
 {
 	const u8 *s = src;
 	u8 *d = dest;
